Refactored structure of auth | Updated imports
