+++ forttask/src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import NextAuth from "next-auth";
+import { authOptions } from "../../../auth";
+
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
