+++ forttask/src/app/api/household/users/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import prisma from '../../../../../libs/prisma';
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '../../../../pages/api/auth/[...nextauth]';
+import { authOptions } from '../../../auth';
 
 export async function GET(req: Request) {
     try {
