+++ forttask/src/app/auth.ts
@@ -1,6 +1,6 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
+import { NextAuthOptions } from 'next-auth';
 import CredentialsProvider from 'next-auth/providers/credentials';
-import prisma from '../../../../libs/prisma';
+import prisma from '../../libs/prisma';
 import bcrypt from 'bcrypt';
 
 export const authOptions: NextAuthOptions = {
@@ -125,5 +125,3 @@ export const authOptions: NextAuthOptions = {
         signIn: '/login',
     },
 };
-
-export default NextAuth(authOptions);
